Resolved layout issues with NSSynchronizationStatusDetailController, moved layoutTableHeaderView method to a separate extension file
