+++ DP3TApp/SharedUI/Helpers/UITableViewExtension.swift
@@ -1,0 +1,39 @@
+//
+/*
+ * Copyright (c) 2020 Ubique Innovation AG <https://www.ubique.ch>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+import Foundation
+
+extension UITableView {
+    func layoutTableHeaderView() {
+        guard let headerView = tableHeaderView else { return }
+        headerView.translatesAutoresizingMaskIntoConstraints = false
+        
+        let headerWidth = headerView.bounds.size.width
+        let temporaryWidthConstraints = NSLayoutConstraint.constraints(withVisualFormat: "[headerView(width)]", options: NSLayoutConstraint.FormatOptions(rawValue: UInt(0)), metrics: ["width": headerWidth], views: ["headerView": headerView])
+        
+        headerView.addConstraints(temporaryWidthConstraints)
+        
+        headerView.setNeedsLayout()
+        headerView.layoutIfNeeded()
+        
+        let headerSize = headerView.systemLayoutSizeFitting(UIView.layoutFittingCompressedSize)
+        let height = headerSize.height
+        var frame = headerView.frame
+        
+        frame.size.height = height
+        headerView.frame = frame
+        
+        tableHeaderView = headerView
+        
+        headerView.removeConstraints(temporaryWidthConstraints)
+        headerView.translatesAutoresizingMaskIntoConstraints = true
+    }
+}
